feat:Retornando usuários que mais compraram

+++ src/main/java/com/foursales/e_commerce/infrastructure/service/ProductServiceImpl.java
@@ -7,15 +7,9 @@ import com.foursales.e_commerce.repository.OrderProductRepository;
 import com.foursales.e_commerce.repository.OrderRepository;
 import com.foursales.e_commerce.repository.ProductRepository;
 import com.foursales.e_commerce.repository.UserRepository;
-import com.foursales.e_commerce.repository.entity.OrderEntity;
-import com.foursales.e_commerce.repository.entity.OrderProductEntity;
-import com.foursales.e_commerce.repository.entity.User;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.UUID;
 
 @Service
 public record ProductServiceImpl(ProductRepository productRepository,
@@ -40,14 +34,30 @@ public record ProductServiceImpl(ProductRepository productRepository,
     }
 
     @Override
-    public Product updateProduct(UUID id, Product product) {
+    public Product updateProduct(Product product) {
+
+        try {
+            var productEntity = productMapper.productToProductEntity(product);
+            productRepository.save(productEntity);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+
         return null;
     }
 
     @Override
     public void deleteProduct(String id) {
+
         try {
-            productRepository.deleteById(id);
+            if (orderProductRepository.existsByProduct_Id(id)) {
+                productRepository.findById(id).ifPresent(produto -> {
+                    produto.setAtivo(false);
+                    productRepository.save(produto);
+                });
+            } else {
+                productRepository.deleteById(id);
+            }
 
         } catch (Exception e) {
             throw new RuntimeException(e);
@@ -66,46 +76,9 @@ public record ProductServiceImpl(ProductRepository productRepository,
 
     @Override
     public Product getProductById(String id) {
-        return productMapper.productEntityToProduct(productRepository.findById(id).get());
-    }
 
-    public Product creatdeProduct(Product product) {
-        var productEntity = productMapper.productToProductEntity(product);
-
-        User user = User.builder()
-                .name("João Silva")
-                .cpf("12345678901")
-                .password("senha123")
-                .creationDate(LocalDateTime.now())
-                .updateDate(LocalDateTime.now())
-                .build();
-
-        OrderEntity orderEntity = OrderEntity.builder()
-                .status("PENDENTE")
-                .value(new BigDecimal("4500.00"))
-                .creationDate(LocalDateTime.now())
-                .updateDate(LocalDateTime.now())
-                .user(user) // Associando o User ao Order
-                .build();
-
-
-        OrderProductEntity orderProduct = OrderProductEntity.builder()
-                .orderEntity(orderEntity) // Associando o Order ao OrderProduct
-                .product(productEntity) // Associando o Product ao OrderProduct
-                .productQuantity(1)
-                .build();
-
-
-        try {
-            userRepository.save(user);
-            orderRepository.save(orderEntity);
-            productRepository.save(productEntity);
-            orderProductRepository.save(orderProduct);
-
-        } catch (Exception e) {
-//            TODO: CRIAR EXCEPTION EXCLUSIVA
-            new RuntimeException(e);
-        }
-        return null;
+        return productRepository.findById(id)
+                .map(productMapper::productEntityToProduct)
+                .orElse(null);
     }
 }
